feat: WritePage 썸네일이미지 등록 버튼 누르면 이미지 등록되도록 수정
