+++ src/Components/Editor.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { CKEditor } from '@ckeditor/ckeditor5-react';
 import ClassicEditor from '@ckeditor/ckeditor5-build-classic';
-import parse from 'html-react-parser';
 
 export default function Editor({
   content,
